refactor: 컴퓨터 랜덤 숫자 생성 기능 함수 파일 분리
